Reorganized the test code and updated the gemspec.
